scraper-python/requirements.txt
playwright
pytesseract
opencv-python
requests
pyyaml
logging
